+++ libntwdk/ntwin32helper.h
@@ -47,3 +47,7 @@ int WinGetErrorMessage(ULONG ErrorCode,PWSTR *ppMessage);
 int WinGetSystemErrorMessage(ULONG ErrorCode,PWSTR *ppMessage);
 void WinFreeErrorMessage(PWSTR pMessage);
 
+PVOID WinLocalAlloc(ULONG Flags, SIZE_T uBytes);
+PVOID WinLocalReAlloc(PVOID pMem,SIZE_T uBytes,ULONG uFlags);
+PVOID WinLocalFree(PVOID pMem);
+SIZE_T WinLocalSize(PVOID pMem);